refactor: Update KeychainShowQR.vue to improve dialog close functionality
REVERT back to version of this from before changes to POS and separate the POS dialog into a new file
KeychainShowQR reverted back to Hash 14a505fb12ed433cf9beee506596305908a0e029

Which was before adding all the KeepSats functionality for the POS page. This broke the functionality of the original.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>