refactor: Fix currency conversion bug in ConvertWrapper.vue
The code changes in ConvertWrapper.vue fix a bug related to currency conversion. The previous implementation was not correctly handling the conversion of values when the currency was set to "sats". This bug caused incorrect display of balances and amounts. The fix ensures that the conversion is done correctly for all currencies, including "sats".

Signed-off-by: Brian of London (home imac) <brian@v4v.app>