feat: Update HiveLogin.vue to enable/disable buttons based on conditions
Remove unnecessary code and refactor HiveLogin.vue to handle the enable/disable logic for the keychain and HAS buttons. The buttons will now be disabled if the hiveAccObj value is undefined, empty, or null, or if the isKeychain value is false. Additionally, the color of the buttons will be set to grey-9 when disabled. This improves the user experience and ensures that the buttons are only clickable when the necessary conditions are met.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>