feat: Add clear local storage functionality
The `clear_local_storage` and `clear_local_storage_message` keys have been added to the English language file. This allows for the implementation of a clear local storage functionality, which will clear all local storage and log the user out. The message provides a confirmation prompt to ensure the user intends to perform this action.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>