chore: Add dataLoading flag to storeUser.js
The `storeUser.js` file was modified to add a `dataLoading` flag to the store. This flag is used to track the loading state of data in the user store. It is set to `true` before making API calls and set to `false` after the API calls are completed. This change improves the user experience by providing visual feedback when data is being loaded.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>