feat: Add lazy loading for images in CreditCard.vue
The CreditCard.vue component was modified to add lazy loading for images. This change improves the page load performance by loading images only when they are visible on the screen. The lazy loading is implemented using the q-spinner component to show a loading indicator while the images are being loaded.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>