chore: Add dataLoading flag and loading indicator to improve user experience
The `storeUser.js` file was modified to add a `dataLoading` flag to track the loading state of data in the user store. This flag is set to `true` before making API calls and set to `false` after the API calls are completed. Additionally, a loading indicator was added to the Refresh button in `HiveLightningKeepSatsTrans.vue` to provide visual feedback when data is being loaded. These changes improve the user experience by providing better feedback and responsiveness.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>