refactor: Update UserList component to conditionally display avatars and icons
The UserList component has been modified to conditionally display avatars and icons based on the user's login type. The avatar and tooltip for the user's Hive account name are now only shown if the login type is 'hive'. Additionally, the Hive Auth logo icon is only displayed if the user is identified as a HAS user, and the Hive Keychain logo icon is only shown if the user is identified as a Keychain user and the login type is 'hive'.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>