refactor: Update HiveInputAcc component to handle EVM addresses
This commit updates the HiveInputAcc component to handle Ethereum Virtual Machine (EVM) addresses. It imports the necessary functions from the useEVM module, including useIsEVMAddress and useShortEVMAddress. When a Hive profile is not found for a given value, the component checks if the value is an EVM address using useIsEVMAddress. If it is, the component updates the modelValue's caption with the short EVM address using useShortEVMAddress. This allows the component to handle EVM addresses in addition to regular Hive addresses.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>