refactor: Refactor useStoreUser and useV4vapp functions
Refactor the useStoreUser function in storeUser.js to handle the case where the useKeepSats function returns an error with the detail "Could not validate credentials". If this error occurs, the user is logged out and the function returns false.

In useV4vapp.js, update the useKeepSats function to handle the same error case. If the error detail is either "Hive Key authorization failure" or "Could not validate credentials", a notification is displayed to the user and the function returns the error response data.

These changes improve the error handling and user experience when validating credentials in the application.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>